requirement.txt
livekit-agents
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
pyautogui
ddgs
pycaw
comtypes
PySide6
spotipy